Remove some warnings

+++ example/mmc3.c
@@ -61,7 +61,7 @@ main(int argc, const char *argv[])
   }
   
   ret = mmc_get_tray_status(p_cdio);
-  switch(ret) {
+  switch((int) ret) {
   case 0:
     printf("CD-ROM drive %s is closed.\n", psz_drive);
     do_eject = true;
@@ -77,7 +77,7 @@ main(int argc, const char *argv[])
   }
   
   ret = mmc_get_media_changed(p_cdio);
-  switch(ret) {
+  switch((int) ret) {
   case 0:
     printf("CD-ROM drive %s media not changed since last test.\n", psz_drive);
     break;
@@ -96,7 +96,7 @@ main(int argc, const char *argv[])
     ret = cdio_close_tray(psz_drive, &driver_id);
 
   ret = mmc_get_tray_status(p_cdio);
-  switch(ret) {
+  switch((int) ret) {
   case 0:
     printf("CD-ROM drive %s is closed.\n", psz_drive);
     break;
@@ -110,7 +110,7 @@ main(int argc, const char *argv[])
   }
   
   ret = mmc_get_media_changed(p_cdio);
-  switch(ret) {
+  switch((int) ret) {
   case 0:
     printf("CD-ROM drive %s media not changed since last test.\n", psz_drive);
     break;
